backend/pyproject.toml
[project]
name = "plAiground_backend"
version = "0.1.0"
description = "Legal AI Research Sandbox Backend - Secure ephemeral environment for legal AI research"
requires-python = ">=3.12"
dependencies = [
    "bcrypt>=4.3.0",
    "dotenv>=0.9.9",
    "fastapi>=0.116.1",
    "httpx>=0.28.1",
    "pdfplumber>=0.11.7",
    "pydantic-settings>=2.10.1",
    "pypdf>=6.1.0",
    "python-docx>=1.2.0",
    "python-jose>=3.5.0",
    "python-multipart>=0.0.20",
    "reportlab>=4.4.4",
    "uvicorn>=0.35.0",
]

[project.optional-dependencies]
dev = []

[tool.uv]
# Ensure uv uses the venv in the project directory
managed = true

[tool.pyright]
include = ["app"]
exclude = ["**/__pycache__", ".venv"]
venvPath = "."
venv = ".venv"
pythonVersion = "3.12"
typeCheckingMode = "basic"
reportMissingImports = "warning"
autoImportCompletions = true

[tool.black]
line-length = 88
target-version = ['py312']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 88
known_first_party = ["app"]
